Refactor import paths and update dependencies
