+++ src/components/request-management/admin/request-management/request-modal.tsx
@@ -7,29 +7,9 @@ import {
   DialogDescription,
 } from '@/components/ui/dialog';
 import { Calendar, FileText, DollarSign } from 'lucide-react';
-import { format } from 'util';
 import { Badge } from '../../../ui/badge';
-
-const formatDate = (dateString: string) =>
-  format(new Date(dateString), 'DD/MM/YYYY');
-
-const getRequestType = (typeId: number) => {
-  const types = {
-    1: 'Vacaciones',
-    2: 'Certificado de Salario',
-    3: 'Confirmación de Pago',
-  };
-  return types[typeId as keyof typeof types] || 'Desconocido';
-};
-
-const getRequestState = (stateId: number) => {
-  const states = {
-    1: 'Pendiente',
-    2: 'Aprobado',
-    3: 'Rechazado',
-  };
-  return states[stateId as keyof typeof states] || 'Desconocido';
-};
+import { formatDate } from '@/lib/utils';
+import { getRequestState, getRequestType } from '../../request-helpers';
 
 const RequestModal = ({
   request,
