+++ src/components/request-management/user/request-management/request-card.tsx
@@ -6,6 +6,7 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
+import { formatDate } from '@/lib/utils';
 import { RequestDetails } from '@/types';
 import { format } from '@formkit/tempo';
 import {
@@ -14,28 +15,7 @@ import {
   DollarSign,
   Clock,
 } from 'lucide-react';
-
-// Helper functions
-const formatDate = (dateString: string) =>
-  format(new Date(dateString), 'DD/MM/YYYY');
-
-const getRequestType = (typeId: number) => {
-  const types = {
-    1: 'Vacaciones',
-    2: 'Certificado de Salario',
-    3: 'Confirmación de Pago',
-  };
-  return types[typeId as keyof typeof types] || 'Desconocido';
-};
-
-const getRequestState = (stateId: number) => {
-  const states = {
-    1: 'Pendiente',
-    2: 'Aprobado',
-    3: 'Rechazado',
-  };
-  return states[stateId as keyof typeof states] || 'Desconocido';
-};
+import { getRequestState, getRequestType } from '../../request-helpers';
 
 const getRequestIcon = (typeId: number) => {
   switch (typeId) {
